requirements-ci.txt
pytest-cov
coveralls
flake8
bddcli
bddrest

